int to unsigned int - comparison

+++ Sender/monthlyTimestamps-test.cpp
@@ -14,10 +14,10 @@ TEST_CASE("monthly-timestamp")
 	std::vector<int> footfalls
 		= { 81,21,37,27,23,98,45,37,31,25,58,104,64,78,20,68,60,73,30,28,13,56,51,56,67,36,77,70,45,68,83 };
 	
-	for (int i = 0; i < footfalls.size(); i++)
+	for (unsigned int i = 0; i < footfalls.size(); i++)
 	{
-		REQUIRE((test.GetHoursOn(i)).size() == footfalls[i]);
-		REQUIRE((test.GetMinutesOn(i)).size() == footfalls[i]);
+		REQUIRE((test.GetHoursOn(i)).size() == (unsigned int)footfalls[i]);
+		REQUIRE((test.GetMinutesOn(i)).size() == (unsigned int)footfalls[i]);
 	}
 }
